Create RandomizeColumnsArraybyNumbersLimeSurvey
INTRODUCTION

Hey again, guys!

As some of you may know, currently LimeSurvey (Cloud Version 3.27.16) does not have an option to randomize columns in an Array by Numbers.
Although, there may be other solutions to overcome this challenge, sometimes a client can be pretty insistent on a particular research design, hence my LimeSurvey quick workaround.

HERE IS AN EXAMPLE

Say, we are given a list of cars a LimeSurvey research participant needs to select in Question 1 (Q1). Our Q1 question states: “Which car would you like to buy and why?”
One of the design requirements is to ensure the selected cars appear in Question 2 (Q2, Y-Scale).
Further, we need to choose statements (X-Scale) associated with our choices made in Q1. Our Question 2 research statements are: “I like the way it looks”, “I like the way it drives”, “The price is within my budget”.
Lately, we need to make sure we randomize the statements presented in columns (i.e. X-Scale) of the Q2 LimeSurvey Array by Numbers.
The out-of-the-box solution offers randomization of Rows and not Columns.

SOLUTION

As usual, all you need to do guys is to click on my GitHub link in the description of this video, get the script and place it in the Expression Manager just like I do it in this tutorial.